backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.4,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=3.2,<4.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
email-validator>=2.0,<3.0
